Fix issue on detecting if legacy code gen was enabled
