+++ gradle.properties
@@ -15,8 +15,7 @@ org.gradle.jvmargs=-Xmx1536m
 # Android operating system, and which are packaged with your app's APK
 # https://developer.android.com/topic/libraries/support-library/androidx-rn
 android.useAndroidX=true
-# Automatically convert third-party libraries to use AndroidX
-android.enableJetifier=true
+
 GROUP=com.careem.mockingbird
 VERSION=2.7.0-SNAPSHOT
 POM_NAME=mockingbird
