+++ mockingbird-compiler/src/main/kotlin/com/careem/mockingbird/MockingbirdPlugin.kt
@@ -17,35 +17,15 @@ package com.careem.mockingbird
 
 import org.gradle.api.Plugin
 import org.gradle.api.Project
-import org.gradle.api.logging.Logger
-import org.gradle.api.logging.Logging
-import org.gradle.kotlin.dsl.add
-import org.gradle.kotlin.dsl.get
-
-internal const val EXTENSION_NAME = "mockingBird"
 
 abstract class MockingbirdPlugin : Plugin<Project> {
 
     private val mockingbirdPluginLegacyCodeGenDelegate = MockingbirdPluginLegacyCodeGenDelegate()
     private val mockingbirdPluginKspDelegate = MockingbirdPluginKspDelegate()
 
-    private val logger: Logger = Logging.getLogger(this::class.java)
-
     override fun apply(target: Project) {
-        target.extensions.add<MockingbirdPluginExtension>(
-            EXTENSION_NAME, MockingbirdPluginExtensionImpl(target.objects)
-        )
-        if (legacyCodeGenRequired(target)) {
-            mockingbirdPluginLegacyCodeGenDelegate.apply(target)
-        }
+        mockingbirdPluginLegacyCodeGenDelegate.apply(target)
         mockingbirdPluginKspDelegate.apply(target)
-    }
-
-    private fun legacyCodeGenRequired(target: Project): Boolean {
-        val pluginExtensions = target.extensions[EXTENSION_NAME] as MockingbirdPluginExtensionImpl
-        val legacyCodeGenRequested = pluginExtensions.generateMocksFor.isNotEmpty()
-        logger.info("LegacyCodeGen: $legacyCodeGenRequested")
-        return legacyCodeGenRequested
     }
 }
